Struct tokio_tls::proto::Client

pub struct Client<T> { /* fields omitted */ }

TLS client protocol wrapper.

This structure is a wrapper for other implementations of ClientProto in the tokio-proto crate. This structure will negotiate a TLS connection first and then delegate all further operations to the ClientProto implementation for the underlying type.

Methods

impl<T> Client<T>

pub fn new(protocol: T, connector: TlsConnector, hostname: &str) -> Client<T>

Constructs a new TLS protocol which will delegate to the underlying protocol specified.

The connector provided will be used to configure the TLS connection. Further I/O will go through the negotiated TLS stream through the protocol specified.