Struct tokio_tls::proto::Server

pub struct Server<T> { /* fields omitted */ }

TLS server protocol wrapper.

This structure is a wrapper for other implementations of ServerProto in the tokio-proto crate. This structure will negotiate a TLS connection first and then delegate all further operations to the ServerProto implementation for the underlying type.

Methods

impl<T> Server<T>

pub fn new(protocol: T, acceptor: TlsAcceptor) -> Server<T>

Constructs a new TLS protocol which will delegate to the underlying protocol specified.

The acceptor provided will be used to accept TLS connections. All new connections will go through the TLS acceptor first and then further I/O will go through the negotiated TLS stream through the protocol specified.