Crate tokio_tls

Async TLS streams

This library is an implementation of TLS streams using the most appropriate system library by default for negotiating the connection. That is, on Windows this library uses SChannel, on OSX it uses SecureTransport, and on other platforms it uses OpenSSL.

Each TLS stream implements the Read and Write traits to interact and interoperate with the rest of the futures I/O ecosystem. Client connections initiated from this crate verify hostnames automatically and by default.

This crate primarily exports this ability through two extension traits, TlsConnectorExt and TlsAcceptorExt. These traits augment the functionality provided by the native-tls crate, on which this crate is built. Configuration of TLS parameters is still primarily done through the native-tls crate.

Structs

AcceptAsync
Future returned from TlsAcceptorExt::accept_async which will resolve once the accept handshake has finished.

ConnectAsync
Future returned from TlsConnectorExt::connect_async which will resolve once the connection handshake has finished.

TlsStream
A wrapper around an underlying raw stream which implements the TLS or SSL protocol.

Traits

TlsAcceptorExt
Extension trait for the TlsAcceptor type in the native_tls crate.

TlsConnectorExt
Extension trait for the TlsConnector type in the native_tls crate.