Crate tokio_tls

Async TLS streams

Note: This crate is deprecated and has been moved into the [tokio-native-tls] crate.

This library is an implementation of TLS streams using the most appropriate system library by default for negotiating the connection. That is, on Windows this library uses SChannel, on OSX it uses SecureTransport, and on other platforms it uses OpenSSL.

Each TLS stream implements the Read and Write traits to interact and interoperate with the rest of the futures I/O ecosystem. Client connections initiated from this crate verify hostnames automatically and by default.

This crate primarily exports this ability through two newtypes, TlsConnector and TlsAcceptor. These newtypes augment the functionality provided by the native-tls crate, on which this crate is built. Configuration of TLS parameters is still primarily done through the native-tls crate.

Structs

TlsAcceptor

A wrapper around a native_tls::TlsAcceptor, providing an async accept method.

TlsConnector

A wrapper around a native_tls::TlsConnector, providing an async connect method.

TlsStream

A wrapper around an underlying raw stream which implements the TLS or SSL protocol.