Struct tokio_tls::TlsStream

pub struct TlsStream<S>(_);

A wrapper around an underlying raw stream which implements the TLS or SSL protocol.

A TlsStream<S> represents a handshake that has been completed successfully and both the server and the client are ready for receiving and sending data. Bytes read from a TlsStream are decrypted from S and bytes written to a TlsStream are encrypted when passing through to S.

Implementations

impl<S> TlsStream<S>

pub fn get_ref(&self) -> &S where
S: AsyncRead + AsyncWrite + Unpin,

Returns a shared reference to the inner stream.

pub fn get_mut(&mut self) -> &mut S where
S: AsyncRead + AsyncWrite + Unpin,

Returns a mutable reference to the inner stream.